Improved performance notably for long comments.
For example, if there is 20000 long C-style comment and we get requested to classify something near the end, so far we always went back through all the previous lines to find the comment start. This is VERY slow, especially because we get one request per line to classify the line. Therefore, we now cache the result of having identified the comment type. Currently, the cache gets cleared after every edit of the file.