More robust implementation of the caching.
Problem was: We use the default Visual Studio cpp colorer to identify whether some piece of text is in a comment and in which type of comment. However, VS first triggers a GetClassificationSpans(), then after some time the VS cpp colorer is updated and triggers another GetClassificationSpans(). In-between, the text buffer version does not change. Hence, we ended up using the result from the first classification (which we cached) also in the second call. But because the VS cpp colorer was updated, the result was wrong. Specifically, pasting some C-style comment sometimes failed to receive the proper comment highlighting.

One way to fix this would have been to remove the caching. However, the caching is reducing the CPU load quite notably, especially for very long C-style comments. So we want to keep it.
Thus, as solution, we listen to the VS cpp colorer triggering the TagsChanged event. It triggers this once it finished updating its internal view of the code. When it triggers, we remove our caches and trigger a reclassify.