Changed GUID for the options page.
In a recent commit, some GUID was added. It was missing before.
In a futile attempt a GUID was taken that probably was the GUID
generated automatically by the .NET framework in the experimental
instance of Visual Studio, or something like that. Not sure. And
because I am not sure what that GUID actually was, I am changing
it to some freshly generated one.
In a few tests it seems to not break the installation when
updating the extension. Let's hope for the best.