First version of using the Doxygen help as source for the autocompletion.
The python script reads the Doxygen's website's html file and generates a C# source file from it, which is then compiled into VSDoxyHighlighter and used to populate the autocomplete box.

The basic thing works, but especially the descriptions need more work. Also, stuff like \param[in] needs to be added. And maybe the sorting must be amended.