Refactoring: Moved the handling of the user-configured Doxygen commands into the DoxygenCommands class and made it a singleton on the package.
Before, every CommentClassifier and associated CommentParser handled the list of Doxygen commands themselves. However, this makes no sense, since there should only be a single "manager". Also, we need to access it from the autocompletion code, and getting the comment parser for this felt like a hack.

So instead now the user-configured commands are handled by a single instance of DoxygenCommands, which lives on the VSDoxyHighlighterPackage. Everything is using that one now.

For the unit tests, it was necessary to introduce a dedicated interface IGeneralOptions to separate the actual options from the GUI code. This allows us to write a fake in the tests.