Decided on how to deal with quick info's async'ness.
As explained in the comments, at least certain parts of the CommentClassifier needs to execute on the main thread. Or alternatively, we use another thread, and rely on the caches in the CommentClassifier (together with appropriate synchronization).

For simplicity, and because the quick info stuff will most likely not hurt the UI responsiveness (once we exploit the caches in the CommentClassifier), we simply switch to the UI thread in GetQuickInfoItemAsync().