Added @{ and @} commands to the autocomplete and quick info boxes.
For some reason, these two commands are no explained on the Doxygen help page with all commands. So, the Python script cannot extract them, and we have to add them manually.

To this end, adapted the Python script to not only include the anchors in the generated C# file but the full link to the online help. This enables us to use the proper link to the documentation page for @{ and @} (which, again, is on a completely different webpage).