Played around with an alternative way to get all classes.
Getting the list of classes via CodeModel can be quite slow (200,000 classes in ~10s).
Via IVsObjectManager2 we also can get the list. This is an interface that is used to populate the "Class View" window in Visual Studio. Result: It is roughly equally slow or maybe a bit slower :-(
I am just commiting this here in case I need the code again later, but most likely I will use CodeModel. Using IVsObjectManager2 is also somewhat a hack: I think it is not meant to be used by extensions, instead it is supposed to be implemented by extensions.