Implemented heuristic to make a sensible preselection for the autocompletion of the parameters of the @param and @tparam commands.
People typically document one function parameter after the next. So if one types "@param" to document the second function parameter after having documented the first function parameter, it is convenient if the autocomplete box already has the second function parameter preselected. The user then just has to hit the tab key to insert the function parameter. This heuristic is implemented in this commit, both for @param and for @tparam, for functions, classes and macros.