Improved autocompletion of parameters when more complicated types are involved.
I.e. when the types of NTTP or function parameters are more complicated. Unfortunately, using the SemanticTokensCache, it seems impossible to find the next "interesting" C++ element reliably, especially to stop at the correct element. For example SemanticTokenKind.cppClassTemplate can appear in both template parameter lists and is also used for the actual class declaration/definition.
The FileCodeModel gives more reliable information, but only if it does give any information at all... none for declarations... So we cannot ditch the SemanticTokensCache.
In the end, we just can implement sensible heuristics and hope that most cases are supported, with not too many wrong results.