Implemented simple heuristic to get the type of a function parameter in the context of the SemanticTokensCache.
We rely on the SemanticTokensCache and cannot use the FileCodeModel for global declarations. This is quite common in C/C++ code. So it makes sense to try and get the parameter information.
The heuristic assumes that the whole type is on the same line as the name and no C-style or string literals are involved. We don't want to write a full C++ parser after all.