Removed incorrect assert.
When typing @param within a function above a function call, the assert got triggered.
We actually now list the parameters of the calling (not the called) function in the autocomplete box. I wanted to change this, but the FileCodeModel's start offset of the element seems wrong. Oh well, showing the autocomplete box doesn't really hurt, so I keep it for now.