Write a version number to the config file, and use it so that we meddle with read configurations only for old versions.
In newer versions of the extension, we need to change the config format every now and then. But we still want to be able to read old configs. To support this, we now write a version number to the config file.

Note: To be precise, we could still do without it. Nevertheless, in the future we might encounter situations where we really need to know the version of the config, lest we make wrong amendments. So implement it now, to prevent future problems.