Use an integer to write the config version.
Reason: The "export settings" to XML feature of Visual Studio seems to write the enum as label. We don't want that, especially considering that ConfigVersions.v1_8_0 and ConfigVersions.Current have the same numeric value.

Note: The settings are stored in the registry. There it already seemed to have used the numeric value.