Suppress nuget warnings of levels moderate and below.
The Microsoft.VisualStudio.SDK package indirectly pulls in MessagePack in an old version. I do not want to update the Microsoft.VisualStudio.SDK package because then the extension won't work in older VS 2022 versions. For example, updating Microsoft.VisualStudio.SDK to 17.12 means that the extension requires VS 2022 17.12, at least from a quick test where it failed to load in VS 2022 17.7.

On the other hand, the new automatic warning about security issues in nuget packages in VS is annoying.

Considering the scope of the extension, <=moderate security warnings should be irrelevant, so disable them.