Preparation: Hide the underlying parsing method (i.e. regex) behind interfaces.
Introduce abstract interfaces between the Doxygen commands and the underlying regex parsers in this commit. Currently, only the regex parsers exist, but I intend to introduce more manual parsers in future commits (namely, for parsing the options of e.g. the \snippet command). Regex is simply not well suited for certain tasks.