Refactoring/clarifications regarding "clamped" options of Doxygen parameters.
A previous commit introduced a more general machinery for parsing the "{...}" of e.g. "\snippet{...}". My intention was to also use it to parse "\param[...]", "\htmlonly[block]" and "\htmlinclude[block]". However, after more experimenting with Doxygen, it turns out that the behavior between "\snippet{...}" and "\param[...]" is very different. Moreover, the behavior of "\param[...]" vs "\htmlonly[block]"/"\htmlinclude[block]" is also different.
Therefore:
- The more general machinery will be used only for braced options "{...}" => Replaced the "clamped" in the names with "braced".
- Will keep the current parsing for "\param" => Added comments and some additional tests.
- Will keep the current parsing for "\htmlonly" and "\htmlinclude" => Added comments.