Fixed crash when Visual Studio decides to load settings multiple times.
E.g. clicking the cancel button in the VS options causes VS to load the settings again, even though the settings shouldn't have changed.
In any case, we need to ensure that the "Version" gets reset before loading the settings, so that we can properly react to old data being read again (just like we did when the settings had been loaded the first time).